Role base login

+++ brandName/app/Http/Controllers/ProductController.php
@@ -1,9 +1,13 @@
 <?php
 
+
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\Authcontroller;
 use App\Products;
 use Illuminate\Http\Request;
+
+use Auth;
 
 class ProductController extends Controller
 {
@@ -12,41 +16,51 @@ class ProductController extends Controller
 
         $products = Products::all();
 
-        return view('products.index',compact('products'));
+            if ( Auth::user()->accout == 1)
+            return view('products.index',compact('products'));
 
+            return view('index');
     }
 
     public function create()
     {
-        return view('products.create');
+        if ( Auth::user()->accout == 1)
+            return view('products.create');
+        return view('index');
     }
 
     public function store(Request $request)
     {
-        Products::create($request->all());
-
-        return redirect()->route('products.index')->with('message','Item has been created successfully!');
-
+        if ( Auth::user()->accout == 1) {
+            Products::create($request->all());
+            return redirect()->route('products.index')->with('message', 'Item has been created successfully!');
+        }
+        return view('index');
     }
 
     public function edit( Products $product )
     {
-        return view('products.edit',compact('product'));
+        if ( Auth::user()->accout == 1)
+            return view('products.edit',compact('product'));
+        return view('index');
     }
 
 
     public function update($id, Request $request)
     {
-        $item = Products::findOrFail($id);
-
-        $item->update($request->all());
-
-        return redirect()->route('products.index')->with('message','Item has been Updated successfully!');
+        if ( Auth::user()->accout == 1) {
+            $item = Products::findOrFail($id);
+            $item->update($request->all());
+            return redirect()->route('products.index')->with('message', 'Item has been Updated successfully!');
+        }
+        return view('index');
     }
 
     public function show(Products $product)
     {
-        return view('products.show',compact('product'));
+        if ( Auth::user()->accout == 1)
+            return view('products.show',compact('product'));
+        return view('index');
     }
 
     /**
@@ -58,14 +72,14 @@ class ProductController extends Controller
     public function destroy($id)
     {
         //$task->delete();
-
-        $item = Products::findOrFail($id);
+        if ( Auth::user()->accout == 1) {
+            $item = Products::findOrFail($id);
 
             $item->delete();
-            return redirect()->route('products.index')->with('message','Item has been Deletd successfully!');
-    }
-    public function home(){
+            return redirect()->route('products.index')->with('message', 'Item has been Deletd successfully!');
+        }
         return view('index');
     }
+
 
 }
